add test/shared.py
