+++ test/test_timeseries.py
@@ -4,24 +4,16 @@ from firebird    import timeseries
 from pyspark     import SparkContext
 from pyspark.sql import SparkSession
 
+from test.shared import acquired
+from test.shared import ard_schema
+from test.shared import aux_schema
+from test.shared import dummy_list
+from test.shared import mock_merlin_create
+from test.shared import mock_timeseries_rdd
+
 import json
 import merlin
 import pyspark.sql.types
-
-def mock_timeseries_rdd(ctx, cids, acquired, cfg, name):
-    rdd = ctx.parallelize([[[11, 22, 33, 44], (-999, 111)]])
-    rdd.setName(name)
-    return rdd
-
-def mock_merlin_create(x, y, acquired, cfg):
-    return [[[11, 22, 33, 44], (x, y)]]
-
-dummy_list = [[5, 4, 3, 2], {"foo": 66}]
-ard_schema = ['chipx', 'chipy', 'x', 'y', 'dates', 'blues', 'greens', 'reds', 'nirs', 'swir1s', 'swir2s', 'thermals', 'qas']
-aux_schema = ['chipx', 'chipy', 'x', 'y', 'dates', 'dem', 'trends', 'aspect', 'posidex', 'slope', 'mpw']
-tile_resp  = json.loads(open("test/data/tile_response.json").read())
-acquired   = '1980-01-01/2017-01-01'
-
 
 def test_schema_invalid():
     assert timeseries.schema("foo") is None
